chore(types): tsify lib/gui/windows and spec
